# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: OpenLane API
  version: 0.1.0
  description: >
    The OpenLane public API — source of truth (ADR-0005).
    Every UI action hits this API. Versioned under /v1; breaking changes need /v2.
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.txt
servers:
  - url: http://localhost:8080/v1
    description: local
paths:
...
  /v1/imports:
    post:
      summary: Import tasks from a competitor CSV (dry-run or commit)
      description: |
        Parses an Asana/Rocketlane/generic CSV export into OpenLane tasks.
        With dry_run=true nothing is written; the response is the full
        per-row validation report. Without it, valid rows are inserted in
        one transaction and invalid rows are quarantined in the response.
      operationId: createImport
      security:
        - bearerAuth: []
      parameters:
        - name: dry_run
          in: query
          schema:
            type: boolean
            default: false
        - name: X-Workspace-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, source]
              properties:
                file:
                  type: string
                  format: binary
                source:
                  type: string
                  enum: [asana, rocketlane, generic]
                project_id:
                  type: string
                  format: uuid
                  description: Existing project to import into (required unless create_project given).
                create_project:
                  type: object
                  properties:
                    customer_id:
                      type: string
                      format: uuid
                    name:
                      type: string
                      minLength: 3
                      maxLength: 120
                  required: [customer_id, name]
      responses:
        "200":
          description: Validation report (dry run) or import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
        "400":
          description: Bad request (unknown source, missing file, no project target)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Target project not found (or out of workspace scope)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
...
  schemas:
...
    ImportResult:
      type: object
      properties:
        dry_run:
          type: boolean
        source:
          type: string
        project_id:
          type:
            - string
            - "null"
          format: uuid
        ok_count:
          type: integer
        errors:
          type: array
          items:
            type: object
            properties:
              source_row:
                type: integer
              reason:
                type: string
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          description: RFC 9457 problem type URI
          examples: [https://openlane.dev/errors/portal-token-expired]
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    portalToken:
      type: apiKey
      in: query
      name: portal_token
security:
  - bearerAuth: []